+++ instantsearch-android-core/src/main/kotlin/com/algolia/instantsearch/core/relevantsort/RelevantSortPresenter.kt
@@ -1,0 +1,6 @@
+package com.algolia.instantsearch.core.relevantsort
+
+/**
+ * Defines the way we want to interact with a Relevant sort priority value.
+ */
+public typealias RelevantSortPresenter<T> = (RelevantSortPriority?) -> T