+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Unreleased
 
+### Added
+- Relevant sort widget (#229)
+
+### Changed
+- Algolia Kotlin Client `1.7.0`
+
 ### Fixed
 - `Subscription`: subscriptions list thread-safety